chore: Create indexes for links table
This commit adds indexes to the 'links' table in the database. Indexes are created for the following columns: 'targetFid', 'fid', 'type', 'fid' and 'targetFid', and 'fid' and 'timestamp' (where 'deleted_at' is null). These indexes improve the performance of queries involving these columns and optimize the retrieval of data from the 'links' table.